#53: initial refactor, heavily WIP
