+++ examples/example-1/jest.config.js
