+++ examples/example-1/project1/app.ts
@@ -1,0 +1,1 @@
+export default (a: number, b: number) => a + b
