+++ examples/example-1/project2/app.spec.ts
@@ -1,0 +1,9 @@
+import app from './app'
+
+describe('project2', () => {
+  it('should run test', async () => {
+    const result = app(1, 2)
+
+    expect(result).toEqual(1 + 2)
+  })
+})
