+++ examples/example-2/.sequelizerc
