+++ examples/example-2/jest.config.js
@@ -5,10 +5,7 @@ module.exports = {
     'ts-jest': {
       diagnostics: false, // https://huafu.github.io/ts-jest/user/config/diagnostics
     },
-    'jest-mock-proxy': {},
   },
   testPathIgnorePatterns: ['/node_modules/'],
   setupFiles: ['./jest.setup.ts'],
-  roots: ['./src'],
-  projects: ['<rootDir>/src', '<rootDir>/example'],
 }
