+++ examples/example-2/jest.setup.ts
@@ -1,0 +1,3 @@
+jest.setTimeout(60 * 1000)
+
+process.env.NODE_ENV = 'test'
