+++ examples/example-2/kafka-1-kafkajs/app.ts
