+++ examples/example-2/package.json
@@ -4,8 +4,7 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "test": "ts-node ./dockest.ts",
-    "test2": "ts-node ./dockest2.ts"
+    "test": "ts-node ./dockest.ts"
   },
   "dependencies": {
     "dotenv": "^8.0.0",
