+++ examples/example-2/postgres-1-sequelize/config/postgresConfig.js
