+++ examples/example-2/postgres-1-sequelize/models/index.ts
