+++ examples/example-2/postgres-2-knex/seeders/20190101001337-demo-banana.js
