+++ examples/example-2/redis-1-ioredis/app.ts
